+++ Tests/FO-DICOM.Benchmark/Program.cs
@@ -9,15 +9,14 @@ namespace FellowOakDicom.Benchmark
 {
     static class Program
     {
-        static void Main()
+        static void Main(string[] args)
         {
-            // Run all benchmarks in assembly
-            BenchmarkRunner.Run(typeof(Program).Assembly,
- //           BenchmarkRunner.Run<ParseDatasetBenchmark>(
-                ManualConfig.Create(DefaultConfig.Instance)
-                .WithOptions(ConfigOptions.JoinSummary)
-                .WithOptions(ConfigOptions.DisableOptimizationsValidator)
-                );
+            // Support --filter arg from command line, otherwise run all benchmarks
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly)
+                .Run(args,
+                    ManualConfig.Create(DefaultConfig.Instance)
+                        .WithOptions(ConfigOptions.JoinSummary)
+                        .WithOptions(ConfigOptions.DisableOptimizationsValidator));
         }
     }
 }